JENA-119 (Eliminate memory bounds during query execution) - made QueryIterDistinct non-memory bound.
git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Jena2/ARQ/trunk@1205673 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/openjena/atlas/data/DistinctDataNet.java
@@ -77,8 +77,11 @@ public class DistinctDataNet<E> extends DistinctDataBag<E>
     protected void deleteSpillFiles()
     {
         super.deleteSpillFiles();
-        FileOps.delete(firstSpillFile, false);
-        firstSpillFile = null;
+        if (null != firstSpillFile)
+        {
+            FileOps.delete(firstSpillFile, false);
+            firstSpillFile = null;
+        }
     }
     
     // Used by the .iterator() method
